adding the cpu temp function

+++ libs/obc_hw_lib/fsw/public_inc/obc_hw_lib.h
@@ -5,7 +5,7 @@
 #include "cfe.h"
 
 
-int32 OBC_HW_LIB_get_cpu_temp(void);
+float OBC_HW_LIB_get_cpu_temp(void);
 
 
 #endif
